lab7/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
python-multipart
redis
pydantic
pytest
httpx
